Consolidated new routes and old components
